requirements.txt
flask
requests
GitPython